# api.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: HackerNews
  description: "This is the Open API documentation for the REST API of our beloved application **HackerNews** deployed at <https://asw.raulplesa.online/>."
servers:
  - url: 'https://asw.raulplesa.online/api'
paths:
...
  /user/{userId}/threads:
    get:
      tags:
      - Users
      summary: Returns the comment threads where the user identified by userId participates
      description: Returns an Array of comments and their respective children
      operationId: threads
      parameters:
      - name: userId
        in: path
        description: id of the user to retrieve its comment threads
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/Comments'
        404:
          description: 'Error: user with userId as ID not found'
...
components:
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        content:
          type: string
          example: 'This is a comment!'
        upvotes:
          type: integer
          example: 381
        authorId:
          type: string
          example: "101908718141570741187"
        authorName:
          type: string
          example: "raul.plesa"
        createdAt:
          type: string
          example: '2021-11-14T20:11:58.057Z'
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
    Comments:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY